backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
pandas==2.2.2
polars==1.6.0
openpyxl==3.1.5
python-dotenv==1.0.1
mysql-connector-python==9.0.0